betting_backend/api/requirements.txt

# Flask API Requirements
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
